Changed script to open, resize and move window each time
This change removes the loop at the end that resizes and moves all windows after they have all been opened. This was not very performant with large numbers of windows. Instead it opens a window and then immediately resizes and moves that window and then continues to the next window.